<feat> 일시정지 패널 캐릭터 컨트롤러로 인풋 받게 변경
<feat> 일시정지 패널 캐릭터 컨트롤러로 인풋 받게 변경
