--- justinbeaver/Assets/01_Scripts/UI/PauseInput.cs
@@ -1,30 +1,0 @@
-﻿using UnityEngine;
-using UnityEngine.InputSystem;
-
-public class PauseInput : MonoBehaviour
-{
-    [SerializeField] private InputActionReference pauseAction;
-
-    private void OnEnable()
-    {
-        if (pauseAction != null)
-        {
-            pauseAction.action.Enable();
-            pauseAction.action.performed += OnPause;
-        }
-    }
-
-    private void OnDisable()
-    {
-        if (pauseAction != null)
-        {
-            pauseAction.action.performed -= OnPause;
-            pauseAction.action.Disable();
-        }
-    }
-
-    private void OnPause(InputAction.CallbackContext ctx)
-    {
-        UIManager.Instance?.TogglePauseUI();
-    }
-}
