--- justinbeaver/Assets/01_Scripts/UI/PauseInput.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 41a3699c03314674aaff304b2f274540
