+++ justinbeaver/Assets/01_Scripts/UI/UIInputController.cs
@@ -128,85 +128,4 @@ public class UIInputController : MonoBehaviour
     //   || (UIManager.Instance.gameSuccessUI != null && UIManager.Instance.gameSuccessUI.activeSelf);
     //}
 
-
-    public void OnBack(InputAction.CallbackContext ctx)
-    {
-        if (!ctx.performed)
-        {
-            return;
-        }    
-            
-        //C키, 뒤로가기
-        if (UIManager.Instance != null && UIManager.Instance.IsPauseOpen)
-        {
-            UIManager.Instance.HandlePauseBack();
-            return;
-        }
-    }
-
-
-    public void OnNavigate(InputAction.CallbackContext ctx)
-    {
-        if (!(ctx.started || ctx.performed))
-        {
-            return;
-        }    
-            
-        if (Time.unscaledTime < nextRepeatTime)
-        {
-            return;
-        }    
-            
-        if (UIManager.Instance == null)
-        {
-            return;
-        }    
-            
-        //도움말 패널이 열려있을 때만 페이지 이동
-        var help = UIManager.Instance.GetHelpPanel();
-        if (help == null || !help.gameObject.activeInHierarchy)
-        {
-            return;
-        }    
-            
-        Vector2 v = ctx.ReadValue<Vector2>();
-
-        //좌, 우만
-        if (v.x > axisThreshold)
-        {
-            help.Next();
-            nextRepeatTime = Time.unscaledTime + repeatDelay;
-        }
-        else if (v.x < -axisThreshold)
-        {
-            help.Prev();
-            nextRepeatTime = Time.unscaledTime + repeatDelay;
-        }
-    }
-
-    public void OnHelpPrev(InputAction.CallbackContext ctx)
-    {
-        if (!ctx.performed)
-        {
-            return;
-        }
-        var help = UIManager.Instance?.GetHelpPanel();
-        if (help != null && help.gameObject.activeInHierarchy)
-        {
-            help.Prev();
-        }
-    }
-
-    public void OnHelpNext(InputAction.CallbackContext ctx)
-    {
-        if (!ctx.performed)
-        {
-            return;
-        }
-        var help = UIManager.Instance?.GetHelpPanel();
-        if (help != null && help.gameObject.activeInHierarchy)
-        {
-            help.Next();
-        }
-    }
 }
