+++ justinbeaver/Assets/03_Scenes/HumanHouseScene.unity
@@ -17062,6 +17062,10 @@ PrefabInstance:
     serializedVersion: 3
     m_TransformParent: {fileID: 1522256523}
     m_Modifications:
+    - target: {fileID: 10845351517468929, guid: 57888d1aa92f82746954c93775089fb4, type: 3}
+      propertyPath: firstMainButton
+      value: 
+      objectReference: {fileID: 5446818436592956870}
     - target: {fileID: 5949769610348630437, guid: 57888d1aa92f82746954c93775089fb4, type: 3}
       propertyPath: m_Pivot.x
       value: 0.5
@@ -17155,6 +17159,17 @@ PrefabInstance:
     m_AddedGameObjects: []
     m_AddedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 57888d1aa92f82746954c93775089fb4, type: 3}
+--- !u!114 &5446818436592956870 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 7228234071773098447, guid: 57888d1aa92f82746954c93775089fb4, type: 3}
+  m_PrefabInstance: {fileID: 5446818436592956869}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1792442052}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 4e29b1a8efbd4b44bb3f3716e73f07ff, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: UnityEngine.UI::UnityEngine.UI.Button
 --- !u!33 &5479218741942437736
 MeshFilter:
   m_ObjectHideFlags: 0
